requirements.txt
appdirs==1.4.4
asgiref==3.3.1
boto3==1.16.30
botocore==1.19.30
certifi==2020.11.8
cfgv==3.2.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-3-jet==1.0.8
django-cors-headers==3.5.0
django-heroku==0.3.1
django-storages==1.10.1
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.10
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
nodeenv==1.5.0
packaging==20.7
Pillow==8.0.1
pre-commit==2.9.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.1
whitenoise==5.2.0
